First Phase of Debug Added for Salt Grains Sync Bug
There appears to be a bug in Salt Grains where the data is not
properly synced during a state.apply, which has caused issues in
Salt States that are dependent upon the Grain data to be watching
for transient interfaces.

My company has an sls file that opens a port-forwarding service
whenever a VPN is open on our systems. The VPN opens and closes
at various times. Whenever a state.apply occurs, (which happens
frequently on our systems... usually about once every hour,
sometimes more often) Salt is expected to enforce this state based
on the Grain ip_interface data.

It has been found through testing whenever a state.apply occurs,
the Grain ip_interface data is not accurate to the actual ip interfaces
on the system. Then, Salt goes through its state.apply execution, and
it appears that the Grain data is refreshed, but whatever data that was
in Grains at the start of the state.apply is what was used when enforcing
the Salt States.

The above behavior causes the first state.apply to not properly enforce
the Salt State, and a second state apply works because the first one
refreshed the data and now the Grain data is properly refreshed at the
start of the second state.apply.

Further testing needs to be done to confirm if this is a bug or not.
If Salt's network.py is changed to check for transient interfaces
and the State object class, located in salt/state.py,  is modified
to always load the Grain data, then this refresh issue doesn't happen.

However, when looking at the above two files, the lines of code that cause
problems appear to be on purpose. That's why I'm trying to really dig deep
and add debug code in every nook and cranny that might be called during
the execution of a state.apply.