Phase 2 of Debug to Fix Grains Sync Issue
This is the `final`(I say this w/ hesitation) phase of debug code
added to follow the callstack of a Salt state.apply.

So far, it seems clear that the Grain data is not properly synced
at the start of a state.apply. More testing and research is to be had.